Change "scale" type to double because it was giving me rounding errors at some point. Eventually, these should be integers.
git-svn-id: file:///srv/local/data/cleanRepo/trunk@16311 a0517364-e873-4167-b1c8-416daa5a6b91

+++ matsim/src/main/java/org/matsim/vis/otfvis/gui/OTFHostControlBar.java
@@ -181,7 +181,7 @@ public final class OTFHostControlBar extends JToolBar implements ActionListener,
 	}
 
 	public void updateScaleLabel() {
-		float scale = 0.0f;
+		double scale = 0.0;
 		for (OTFDrawer drawer : drawers) {
 			if (!(drawer instanceof OTFTimeLine)) {
 				scale = drawer.getScale();
